+++ lib/ogr/feature_definition.rb
@@ -5,6 +5,13 @@ require_relative '../ogr'
 
 module OGR
   class FeatureDefinition
+    # @param pointer [FFI::Pointer]
+    def self.release(pointer)
+      return unless pointer && !pointer.null?
+
+      FFI::OGR::API.OGR_FD_Release(pointer)
+    end
+
     # @return [FFI::Pointer] C pointer of the C FeatureDefn.
     attr_reader :c_pointer
 
@@ -13,21 +20,23 @@ module OGR
     #   FFI::Pointer, the new object will simply wrap the C FeatureDefinition
     #   at that address.
     def initialize(name_or_pointer)
-      @c_pointer = if name_or_pointer.is_a? String
-                     FFI::OGR::API.OGR_FD_Create(name_or_pointer)
-                   else
-                     name_or_pointer
-                   end
+      pointer = if name_or_pointer.is_a? String
+                  FFI::OGR::API.OGR_FD_Create(name_or_pointer)
+                else
+                  name_or_pointer
+                end
 
-      return if @c_pointer.is_a?(FFI::Pointer) && !@c_pointer.null?
+      if !pointer.is_a?(FFI::Pointer) || pointer.null?
+        raise OGR::InvalidFeatureDefinition, "Unable to create #{self.class.name} from #{name_or_pointer}"
+      end
 
-      raise OGR::InvalidFeatureDefinition, "Unable to create #{self.class.name} from #{name_or_pointer}"
+      @c_pointer = FFI::AutoPointer.new(pointer, FeatureDefinition.method(:release))
+      @c_pointer.autorelease = false
     end
 
     def release!
-      return unless @c_pointer
+      FeatureDefinition.release(@c_pointer)
 
-      FFI::OGR::API.OGR_FD_Release(@c_pointer)
       @c_pointer = nil
     end
 
@@ -44,8 +53,9 @@ module OGR
     # @param index [Integer]
     # @return [OGR::FieldDefinition]
     def field_definition(index)
-      field_definition_ptr =
-        FFI::OGR::API.OGR_FD_GetFieldDefn(@c_pointer, index)
+      field_definition_ptr = FFI::OGR::API.OGR_FD_GetFieldDefn(@c_pointer, index)
+      field_definition_ptr.autorelease = false
+
       return nil if field_definition_ptr.null?
 
       OGR::FieldDefinition.new(field_definition_ptr, nil)