+++ lib/ogr/style_table.rb
@@ -5,20 +5,29 @@ require_relative '../ogr'
 
 module OGR
   class StyleTable
+    # @param pointer [FFI::Pointer]
+    def self.release(pointer)
+      return unless pointer && !pointer.null?
+
+      FFI::OGR::API.OGR_STBL_Destroy(pointer)
+    end
+
     # @return [FFI::Pointer] C pointer to the C style table.
     attr_reader :c_pointer
 
     # @param [FFI::Pointer] c_pointer
     def initialize(c_pointer = nil)
-      @c_pointer = c_pointer || FFI::OGR::API.OGR_STBL_Create
+      pointer = c_pointer || FFI::OGR::API.OGR_STBL_Create
+      pointer.autorelease = false
 
-      raise 'Unable to create StyleTable' if @c_pointer.null?
+      raise 'Unable to create StyleTable' if pointer.null?
+
+      @c_pointer = FFI::AutoPointer.new(pointer, StyleTable.method(:release))
     end
 
     def destroy!
-      return unless @c_pointer
+      StyleTable.release(@c_pointer)
 
-      FFI::OGR::API.OGR_STBL_Destroy(@c_pointer)
       @c_pointer = nil
     end